Changes: Db design, Endpoints Featueres: CourseDetail Modal added

+++ UniversityWebApp/DataAccess/Interfaces/ICourseRepository.cs
@@ -5,6 +5,6 @@ namespace UniversityWebApp.DataAccess.Interfaces
     public interface ICourseRepository : IRepository<Course>
     {
         Task<IEnumerable<Course>> GetAll(bool includeStudents);
-        Task<Course?> GetById(ushort courseId);
+        Task<Course?> GetById(int courseId);
     }
 }
