+++ UniversityWebApp/Model/Major.cs
@@ -4,7 +4,7 @@ namespace UniversityWebApp.Model
 {
     public class Major
     {
-        public ushort MajorId { get; set; }
+        public int MajorId { get; set; }
         public string Title { get; set; }
         public ICollection<Student> Students { get; private set; } = new List<Student>();
         public ICollection<CourseTopics> Topics { get; private set; } = new List<CourseTopics>();
